move gasScheduleNotifier mock to testscommon, cleanup duplicated mocks
