+++ testscommon/gasScheduleNotifierMock.go
@@ -1,4 +1,4 @@
-package mock
+package testscommon
 
 import "github.com/ElrondNetwork/elrond-go-core/core"
